feat(traffic_simulator): split hdmap_utils test
